==> R/2_fit_real_data.R <==
# this script fits the catalytic models to the real data

# it depends on:
  #  1. the processed data generated by ~/R/1_process_raw_data.R 
  #  2. the stan models in ~/catalytic-stan-models
  #  3. test sensitivity and specificity specified in ~/utils.R as
  #           a. TEST_SPEC_SENS_1 (default)
  #           b. TEST_SPEC_SENS_2 (alternative)
  #  4. the rstan package

# it generates:
  #  1. a stan-fit object for each of the 4 catalytic models for both the default and 
  #     alternative assumptions about test sensitivity and specificity in
  #           a. ~/fits/real_data/sens_spec_1/ NOTE these are too big to store on the github repo
  #           b. ~/fits/real/data/sens_spec_2/ NOTE these are too big to store on the github repo

source("dependencies.R")

## read in the data

data_sero <- readRDS("data/real/data_sero.rds") # dataframe STUDY * VARIABLES
SEROPOS <- readRDS("data/real/SEROPOS.rds") # STUDY X AGE class matrix of seroprevalence
AGE_L <- readRDS("data/real/AGE_L.rds") # STUDY X AGE class matrix of lower age bound
AGE_U <- readRDS("data/real/AGE_U.rds") # STUDY X AGE class matrix of upper age bound
N_CAMELS <- readRDS("data/real/N_CAMELS.rds") # STUDY X AGE class matrix of N


## pull test sensitivity and specificity

STUDY_TEST_TYPE <- unique(data_sero %>% dplyr::select(STUDY_COUNTRY, TEST_TYPE))
STUDY_TEST_TYPE$id <- 1:nrow(STUDY_TEST_TYPE)
STUDY_TEST_TYPE_1 <- merge(STUDY_TEST_TYPE, TEST_SPEC_SENS_1)
STUDY_TEST_TYPE_1 <- STUDY_TEST_TYPE_1[order(STUDY_TEST_TYPE_1$id),]
STUDY_TEST_TYPE_2 <- merge(STUDY_TEST_TYPE, TEST_SPEC_SENS_2)
STUDY_TEST_TYPE_2 <- STUDY_TEST_TYPE_2[order(STUDY_TEST_TYPE_2$id),]


#################
## EXPONENTIAL ##
#################
#########################
# default sens and spec #
#########################

# seroconversion only

fit1bb_exp_real <- stan(
  file = here::here("catalytic-stan-models/model1bb_exp.stan"),
  data = list(
    S = nrow(SEROPOS),
    A =  ncol(SEROPOS),
    N = N_CAMELS,
    pos = SEROPOS,
    age1 = AGE_L,
    age2 = AGE_U,
    sens = STUDY_TEST_TYPE_1$SENS,
    spec = STUDY_TEST_TYPE_1$SPEC,
    sigma_r = 0,
    sigma_m = 12,
    mabs = -1,
    mu_0 = mu_0_cat,
    mu = mu_cat),
  chains = 4,
  iter = 100000,
  warmup = 5000,
  verbose = TRUE
)

saveRDS(fit1bb_exp_real, "fits/real_data/sens_spec_1/exp/fit1bb.rds")
fit1_summary <- rstan::summary(fit1bb_exp_real)

# seroconversion + sero-reversion

fit2bb_exp_real <- stan(
  file = here::here("catalytic-stan-models/model2bb_exp.stan"),
  data = list(
    S = nrow(SEROPOS),
    A =  ncol(SEROPOS),
    N = N_CAMELS,
    pos = SEROPOS,
    age1 = AGE_L,
    age2 = AGE_U,
    sens = STUDY_TEST_TYPE_1$SENS,
    spec = STUDY_TEST_TYPE_1$SPEC,
    sigma_m = 12,
    mabs = -1,
    mu_0 = mu_0_cat,
    mu = mu_cat),
  chains = 4,
  iter = 100000,
  warmup = 5000,
  verbose = TRUE
)

saveRDS(fit2bb_exp_real, "fits/real_data/sens_spec_1/exp/fit2bb.rds")
fit2_summary <- rstan::summary(fit2bb_exp_real)

# seroconversion + mAbs

fit3bb_exp_real <- stan(
  file = here::here("catalytic-stan-models/model3bb_exp.stan"),
  data = list(
    S = nrow(SEROPOS),
    A =  ncol(SEROPOS),
    N = N_CAMELS,
    pos = SEROPOS,
    age1 = AGE_L,
    age2 = AGE_U,
    sens = STUDY_TEST_TYPE_1$SENS,
    spec = STUDY_TEST_TYPE_1$SPEC,
    sigma_r = 0,
    mabs = 1,
    mu_0 = mu_0_cat,
    mu = mu_cat),
  chains = 4,
  iter = 100000,
  warmup = 5000,
  verbose = TRUE
)

saveRDS(fit3bb_exp_real, "fits/real_data/sens_spec_1/exp/fit3bb.rds")
fit3_summary <- rstan::summary(fit3bb_exp_real)


# seroconversion + mAbs + sero-reversion

fit4bb_exp_real <- stan(
  file = here::here("catalytic-stan-models/model4bb_exp.stan"),
  data = list(
    S = nrow(SEROPOS),
    A =  ncol(SEROPOS),
    N = N_CAMELS,
    pos = SEROPOS,
    age1 = AGE_L,
    age2 = AGE_U,
    sens = STUDY_TEST_TYPE_1$SENS,
    spec = STUDY_TEST_TYPE_1$SPEC,
    mabs = 1,
    mu_0 = mu_0_cat,
    mu = mu_cat
    # B = length(age1_fine),
    # age1_fine = age1_fine,
    # age2_fine = age2_fine
  ),
  chains = 4,
  iter = 10000,
  warmup = 5000,
  verbose = TRUE
)

saveRDS(fit4bb_exp_real, "fits/real_data/sens_spec_1/exp/fit4bb_fine.rds")
fit4_summary <- rstan::summary(fit4bb_exp_real)


#################
## EXPONENTIAL ##
#################
###########################
# alternate sens and spec #
###########################

fit1bb_exp_real_2 <- stan(
  file = here::here("catalytic-stan-models/model1bb_exp.stan"),
  data = list(
    S = nrow(SEROPOS),
    A =  ncol(SEROPOS),
    N = N_CAMELS,
    pos = SEROPOS,
    age1 = AGE_L,
    age2 = AGE_U,
    sens = STUDY_TEST_TYPE_2$SENS,
    spec = STUDY_TEST_TYPE_2$SPEC,
    sigma_r = 0,
    sigma_m = 12,
    mabs = -1,
    mu_0 = mu_0_cat,
    mu = mu_cat),
  chains = 4,
  iter = 100000,
  warmup = 5000,
  verbose = TRUE
)

saveRDS(fit1bb_exp_real_2, "fits/real_data/sens_spec_2/exp/fit1bb.rds")
fit1_summary <- rstan::summary(fit_1bbexp_real_2)

# seroconversion + sero-reversion 

fit2bb_exp_real_2 <- stan(
  file = here::here("catalytic-stan-models/model2bb_exp.stan"),
  data = list(
    S = nrow(SEROPOS),
    A =  ncol(SEROPOS),
    N = N_CAMELS,
    pos = SEROPOS,
    age1 = AGE_L,
    age2 = AGE_U,
    sens = STUDY_TEST_TYPE_2$SENS,
    spec = STUDY_TEST_TYPE_2$SPEC,
    sigma_m = 12,
    mabs = -1,
    mu_0 = mu_0_cat,
    mu = mu_cat),
  chains = 4,
  iter = 100000,
  warmup = 5000,
  verbose = TRUE
)

saveRDS(fit2bb_exp_real_2, "fits/real_data/sens_spec_2/exp/fit2bb.rds")
fit2_summary <- rstan::summary(fit_2bbexp_real)

# seroconversion + mAbs 

fit3bb_exp_real_2 <- stan(
  file = here::here("catalytic-stan-models/model3bb_exp.stan"),
  data = list(
    S = nrow(SEROPOS),
    A =  ncol(SEROPOS),
    N = N_CAMELS,
    pos = SEROPOS,
    age1 = AGE_L,
    age2 = AGE_U,
    sens = STUDY_TEST_TYPE_2$SENS,
    spec = STUDY_TEST_TYPE_2$SPEC,
    sigma_r = 0,
    mabs = 1,
    mu_0 = mu_0_cat,
    mu = mu_cat),
  chains = 4,
  iter = 100000,
  warmup = 5000,
  verbose = TRUE
)

saveRDS(fit3bb_exp_real_2, "fits/real_data/sens_spec_2/exp/fit3bb.rds")
fit3_summary <- rstan::summary(fit3bb_exp_real)


# seroconversion + mAbs + sero-reversion 

fit4bb_exp_real_2 <- stan(
  file = here::here("catalytic-stan-models/model4bb_exp.stan"),
  data = list(
    S = nrow(SEROPOS),
    A =  ncol(SEROPOS),
    N = N_CAMELS,
    pos = SEROPOS,
    age1 = AGE_L,
    age2 = AGE_U,
    sens = STUDY_TEST_TYPE_2$SENS,
    spec = STUDY_TEST_TYPE_2$SPEC,
    mabs = 1,
    mu_0 = mu_0_cat,
    mu = mu_cat),
  #B = length(age1_fine),
  #age1_fine = age1_fine,
  #age2_fine = age2_fine),
  chains = 4,
  iter = 100000,
  warmup = 5000,
  verbose = TRUE
  ##control = list(adapt_delta = 0.99)
)

saveRDS(fit4bb_exp_real_2, "fits/real_data/sens_spec_2/exp/fit4bb.rds")
fit4_summary <- rstan::summary(fit4bb_exp_real_2)

##############################
## UNIFORM AGE DISTIRBUTION ##
##############################

# fit seroconversion
fit1bb_uni_real <- stan(
  file = here::here("catalytic-stan-models/model1bb_uni.stan"),
  data = list(
    S = nrow(SEROPOS),
    A =  ncol(SEROPOS),
    N = N_CAMELS,
    pos = SEROPOS,
    age1 = AGE_L,
    age2 = AGE_U,
    sigma_m = 12,
    sigma_r = 0,
    sens = STUDY_TEST_TYPE_1$SENS,
    spec = STUDY_TEST_TYPE_1$SPEC,
    mabs = -1
    # B = length(age1_fine),
    # age1_fine = age1_fine,
    # age2_fine = age2_fine
  ),
  chains = 4,
  iter = 10000,
  warmup = 2000,
  verbose = TRUE#,
  ##init = init_fun
  ##control = list(adapt_delta = 0.99)
)
saveRDS(fit1bb_uni_real, "fits/real_data/sens_spec_1/uni/fit1bb.rds")

# fit seroconversion + sero-reversion
fit2bb_uni_real <- stan(
  file = here::here("catalytic-stan-models/model2bb_uni.stan"),
  data = list(
    S = nrow(SEROPOS),
    A =  ncol(SEROPOS),
    N = N_CAMELS,
    pos = SEROPOS,
    age1 = AGE_L,
    age2 = AGE_U,
    sigma_m = 12,
    sens = STUDY_TEST_TYPE_1$SENS,
    spec = STUDY_TEST_TYPE_1$SPEC,
    mabs = -1
    # B = length(age1_fine),
    # age1_fine = age1_fine,
    # age2_fine = age2_fine
  ),
  chains = 4,
  iter = 10000,
  warmup = 2000,
  verbose = TRUE
  #control = list(adapt_delta = 0.99) 
)
saveRDS(fit_4_2bbreal, "fits/real_data/sens_spec_1/uni/fit2bb.rds")


# fit seroconversion + mAbs
fit3bb_uni_real <- stan(
  file = here::here("catalytic-stan-models/model3bb_uni.stan"),
  data = list(
    S = nrow(SEROPOS),
    A =  ncol(SEROPOS),
    N = N_CAMELS,
    pos = SEROPOS,
    age1 = AGE_L,
    age2 = AGE_U,
    sigma_r = 0,
    sens = STUDY_TEST_TYPE_1$SENS,
    spec = STUDY_TEST_TYPE_1$SPEC,
    mabs = 1,
    # B = length(age1_fine),
    # age1_fine = age1_fine,
    # age2_fine = age2_fine
  ),
  chains = 4,
  iter = 10000,
  warmup = 2000,
  verbose = TRUE,
  ##control = list(max_treedepth = 15)
  ##control = list(adapt_delta = 0.99)
)
saveRDS(fit3bb_uni_real, "fits/real_data/sens_spec_1/uni/fit3bb.rds")

# fit seroconversion + mAbs + sero-reversion
fit4bb_uni_real <- stan(
  file = here::here("catalytic-stan-models/model4bb_uni.stan"),
  data = list(
    S = nrow(SEROPOS),
    A =  ncol(SEROPOS),
    N = N_CAMELS,
    pos = SEROPOS,
    age1 = AGE_L,
    age2 = AGE_U,
    sens = STUDY_TEST_TYPE_1$SENS,
    spec = STUDY_TEST_TYPE_1$SPEC,
    mabs = 1#,
    # B = length(age1_fine),
    # age1_fine = age1_fine,
    # age2_fine = age2_fine
  ),
  chains = 4,
  iter = 10000,
  warmup = 2000,
  verbose = TRUE
  ##control = list(adapt_delta = 0.99)
)

saveRDS(fit4bb_uni_real, "fits/real_data/sens_spec_1/uni/fit4bb.rds")

# fit seroconversion + mAbs + sero-reversion with og sens and spec
fit4bb_uni_real <- stan(
  file = here::here("catalytic-stan-models/model4bb_uni.stan"),
  data = list(
    S = nrow(SEROPOS),
    A =  ncol(SEROPOS),
    N = N_CAMELS,
    pos = SEROPOS,
    age1 = AGE_L,
    age2 = AGE_U,
    sens = rep(0.999, 23),
    spec = rep(1, 23),
    mabs = 1#,
    # B = length(age1_fine),
    # age1_fine = age1_fine,
    # age2_fine = age2_fine
  ),
  chains = 4,
  iter = 10000,
  warmup = 2000,
  verbose = TRUE
  ##control = list(adapt_delta = 0.99)
)

saveRDS(fit4bb_uni_real, "fits/real_data/sens_spec_1/uni/fit4bb_og_ss.rds")


###############################
## ALTERNATIVE SENS AND SPEC ##
###############################

# seroconversion
fit1bb_uni_real_2 <- stan(
  file = here::here("catalytic-stan-models/model1bb_uni.stan"),
  data = list(
    S = nrow(SEROPOS),
    A =  ncol(SEROPOS),
    N = N_CAMELS,
    pos = SEROPOS,
    age1 = AGE_L,
    age2 = AGE_U,
    sigma_m = 12,
    sigma_r = 0,
    sens = STUDY_TEST_TYPE_2$SENS,
    spec = STUDY_TEST_TYPE_2$SPEC,
    mabs = -1,
    B = length(age1_fine),
    age1_fine = age1_fine,
    age2_fine = age2_fine
  ),
  chains = 4,
  iter = 10000,
  warmup = 3000,
  verbose = TRUE#,
  ##init = init_fun
  ##control = list(adapt_delta = 0.99)
)
saveRDS(fit1bb_uni_real_2, "fits/real_data/sens_spec_2/uni/fit1bb.rds")


# seroconversion + sero-reversion 
fit2bb_uni_real_2 <- stan(
  file = here::here("catalytic-stan-models/model2bb_uni.stan"),
  data = list(
    S = nrow(SEROPOS),
    A =  ncol(SEROPOS),
    N = N_CAMELS,
    pos = SEROPOS,
    age1 = AGE_L,
    age2 = AGE_U,
    sigma_m = 12,
    sens = STUDY_TEST_TYPE_2$SENS,
    spec = STUDY_TEST_TYPE_2$SPEC,
    mabs = -1,
    B = length(age1_fine),
    age1_fine = age1_fine,
    age2_fine = age2_fine
  ),
  chains = 4,
  iter = 10000,
  warmup = 3000,
  verbose = TRUE
  #control = list(adapt_delta = 0.99) 
)
saveRDS(fit2bb_uni_real_2, "fits/real_data/sens_spec_2/uni/fit2bb.rds")


# seroconversion + mAbs 
fit3bb_uni_real_2 <- stan(
  file = here::here("catalytic-stan-models/model3bb_uni.stan"),
  data = list(
    S = nrow(SEROPOS),
    A =  ncol(SEROPOS),
    N = N_CAMELS,
    pos = SEROPOS,
    age1 = AGE_L,
    age2 = AGE_U,
    sigma_r = 0,
    sens = STUDY_TEST_TYPE_2$SENS,
    spec = STUDY_TEST_TYPE_2$SPEC,
    mabs = 1,
    B = length(age1_fine),
    age1_fine = age1_fine,
    age2_fine = age2_fine
  ),
  chains = 4,
  iter = 10000,
  warmup = 3000,
  verbose = TRUE,
  ##control = list(max_treedepth = 15)
  ##control = list(adapt_delta = 0.99)
)
saveRDS(fit3bb_uni_real_2, "fits/real_data/sens_spec_2/uni/fit3bb.rds")

# seroconversion + mAbs + sero-reversion 
fit4bb_uni_real_2 <- stan(
  file = here::here("catalytic-stan-models/model4bb_uni.stan"),
  data = list(
    S = nrow(SEROPOS),
    A =  ncol(SEROPOS),
    N = N_CAMELS,
    pos = SEROPOS,
    age1 = AGE_L,
    age2 = AGE_U,
    sens = STUDY_TEST_TYPE_2$SENS,
    spec = STUDY_TEST_TYPE_2$SPEC,
    mabs = 1,
    B = length(age1_fine),
    age1_fine = age1_fine,
    age2_fine = age2_fine
  ),
  chains = 4,
  iter = 10000,
  warmup = 3000,
  verbose = TRUE
  ##control = list(adapt_delta = 0.99)
)

saveRDS(fit4bb_uni_real_2, "fits/real_data/sens_spec_2/uni/fit4bb.rds")
