==> R/3_process_fits.R <==
# This script processes the raw stan-fits and saves the summarised parameter estimates

# it depends on:
#  1. the stan-fits generated by ~/R/2_fit_real_data.R  and saved in ~/fits/real_data/
#  2. the rstan package

# it generates:
#  1. the summarised parameter estimates for each of the 4 catalytic models
#     under both the default and alternative assumptions about test sensitivity and specificity in:
#       a. fits/processed_real/exp/sens_spec_1/
#       b. fits/processed_real/exp/sens_spec_2/

source("dependencies.R")

#############
# real data #
#############

# read in raw fits

fit1bb_exp_1 <- readRDS("fits/real_data/sens_spec_1/exp/fit1bb.rds")
fit2bb_exp_1 <- readRDS("fits/real_data/sens_spec_1/exp/fit2bb.rds")
fit3bb_exp_1 <- readRDS("fits/real_data/sens_spec_1/exp/fit3bb.rds")
fit4bb_exp_1 <- readRDS("fits/real_data/sens_spec_1/exp/fit4bb.rds")

fit1bb_exp_2 <- readRDS("fits/real_data/sens_spec_2/exp/fit1bb.rds")
fit2bb_exp_2 <- readRDS("fits/real_data/sens_spec_2/exp/fit2bb.rds")
fit3bb_exp_2 <- readRDS("fits/real_data/sens_spec_2/exp/fit3bb.rds")
fit4bb_exp_2 <- readRDS("fits/real_data/sens_spec_2/exp/fit4bb.rds")


fit4bb_exp_1_f <- readRDS("fits/real_data/sens_spec_1/exp/fit4bb_fine.rds")


# summarise

fit1_sum <- gen_summary(fit1bb_exp_1)
fit2_sum <- gen_summary(fit2bb_exp_1)
fit3_sum <- gen_summary(fit3bb_exp_1)
fit4_sum <- gen_summary(fit4bb_exp_1)
fit4_sum_f <- gen_summary(fit4bb_exp_1_f)

fit1_sum2 <- gen_summary(fit1bb_exp_2)
fit2_sum2 <- gen_summary(fit2bb_exp_2)
fit3_sum2 <- gen_summary(fit3bb_exp_2)
fit4_sum2 <- gen_summary(fit4bb_exp_2)

# save summarised parameter estimates

saveRDS(file = "fits/processed_real/exp/sens_spec_1/fit1bb.rds", fit1_sum)
saveRDS(file = "fits/processed_real/exp/sens_spec_1/fit2bb.rds", fit2_sum)
saveRDS(file = "fits/processed_real/exp/sens_spec_1/fit3bb.rds", fit3_sum)
saveRDS(file = "fits/processed_real/exp/sens_spec_1/fit4bb.rds", fit4_sum)
saveRDS(file = "fits/processed_real/exp/sens_spec_1/fit4bb_f.rds", fit4_sum_f)
saveRDS(file = "fits/processed_real/exp/sens_spec_2/fit1bb.rds", fit1_sum2)
saveRDS(file = "fits/processed_real/exp/sens_spec_2/fit2bb.rds", fit2_sum2)
saveRDS(file = "fits/processed_real/exp/sens_spec_2/fit3bb.rds", fit3_sum2)
saveRDS(file = "fits/processed_real/exp/sens_spec_2/fit4bb.rds", fit4_sum2)

###################
# real data - LOO #
###################

# read in raw files

fit1_loo <- readRDS("fits/real_data/sens_spec_1/exp/LOO/1bb_fit.rds")
fit2_loo <- readRDS("fits/real_data/sens_spec_1/exp/LOO/2bb_fit.rds")
fit3_loo <- readRDS("fits/real_data/sens_spec_1/exp/LOO/3bb_fit.rds")
fit4_loo <- readRDS("fits/real_data/sens_spec_1/exp/LOO/4bb_fit.rds")

# summarise

fit1_looSum <- lapply(fit1_loo, FUN = gen_summary)
fit2_looSum <- lapply(fit2_loo, FUN = gen_summary)
fit3_looSum <- lapply(fit3_loo, FUN = gen_summary)
fit4_looSum <- lapply(fit4_loo, FUN = gen_summary)

# save summarised parameter estimates

saveRDS(file = "fits/processed_real/exp/sens_spec_1/fit1_loo.rds", fit1_looSum)
saveRDS(file = "fits/processed_real/exp/sens_spec_1/fit2_loo.rds", fit2_looSum)
saveRDS(file = "fits/processed_real/exp/sens_spec_1/fit3_loo.rds", fit3_looSum)
saveRDS(file = "fits/processed_real/exp/sens_spec_1/fit4_loo.rds", fit4_looSum)
